requirements/dev-3.7.txt
appdirs==1.4.3            # via black
appnope==0.1.0            # via ipython
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, pytest
backcall==0.1.0           # via ipython
black==19.3b0
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via black
coverage==4.5.3           # via coveralls, pytest-cov
coveralls==1.8.1
decorator==4.4.0          # via ipython, traitlets
docopt==0.6.2             # via coveralls
entrypoints==0.3          # via flake8
flake8==3.7.8
google-api-core[grpc]==1.14.0  # via google-cloud-core, google-cloud-kms
google-api-python-client==1.7.10
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3        # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.0.2  # via google-cloud-storage
google-cloud-kms==1.1.0
google-cloud-storage==1.17.0
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos[grpc]==1.6.0  # via google-api-core, grpc-google-iam-v1
grpc-google-iam-v1==0.11.4  # via google-cloud-kms
grpcio==1.22.0            # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
httplib2==0.13.0          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
importlib-metadata==0.18  # via pluggy, pytest
ipdb==0.12.1
ipython-genutils==0.2.0   # via traitlets
ipython==7.6.1            # via ipdb
jedi==0.14.1              # via ipython
mccabe==0.6.1             # via flake8
more-itertools==7.2.0     # via pytest
mypy-extensions==0.4.1    # via mypy
mypy==0.720
packaging==19.0           # via pytest
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.12.0            # via pytest
prompt-toolkit==2.0.9     # via ipython
protobuf==3.9.0           # via google-api-core, googleapis-common-protos
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pycodestyle==2.5.0        # via flake8
pycryptodome==3.8.2
pyflakes==2.1.1           # via flake8
pygments==2.4.2           # via ipython
pyparsing==2.4.1.1        # via packaging
pytest-cov==2.7.1
pytest==5.0.1             # via pytest-cov
pytz==2019.1              # via google-api-core
requests==2.22.0          # via coveralls, google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, packaging, prompt-toolkit, protobuf, traitlets
toml==0.10.0              # via black
traitlets==4.3.2          # via ipython
typed-ast==1.4.0          # via mypy
typing-extensions==3.7.4  # via mypy
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via prompt-toolkit, pytest
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via google-api-core, ipdb, ipython, protobuf
